integer player

+++ app_player_adapter.hpp
@@ -78,8 +78,8 @@ private:
 private:
 	kb::ModPlayer& player_;
 	RingIndex<4096> rw_;
-	std::int16_t buf_[4096*2];
-	float pbuf_[4096*2]; };
+	std::int32_t buf_[4096*2];
+	int pbuf_[4096*2]; };
 
 
 }  // namespace app
